Move from using wwn to serial
We can gather the serial number from lshw, which is a standard linux
tool. While the way that the serial is exposed from lshw may be
different than how the RAID controller may expose it, this should be
good enough. Similarly, this method allows us to use this method on
boxes without a RAID device.