Fix broken ring syncing
We had been using rsync to copy the rings from the local host to the
other nodes. However, this would warn about accepting the foreign host
key, hence non-interactive. Change this to use fabric's put operation
and with it make sure we are doing all of our building in a tempdir.